+++ packages/common/web/src/index.ts
@@ -83,3 +83,4 @@ export { isHexColor } from './Utils/colorUtils';
 export { isRtl, getLangDir } from './Utils/rtlUtils';
 
 export { simplePubsub, Pubsub, Store } from './Utils/simplePubsub';
+export { generateKey } from './Utils/generateKey';