+++ packages/plugin-commons/web/src/Base/createBaseComponent.jsx
@@ -44,6 +44,8 @@ const createBaseComponent = ({
   renderInnerRCE,
   noPluginBorder,
   noPointerEventsOnFocus,
+  withHorizontalScroll,
+  innerRCERenderedIn,
 }) => {
   return class WrappedComponent extends Component {
     static propTypes = {
@@ -269,6 +271,7 @@ const createBaseComponent = ({
       event.dataTransfer.setData('url', this.url || window?.location?.href);
     };
 
+    // eslint-disable-next-line complexity
     render = () => {
       const { blockProps, className, selection } = this.props;
       const { componentData } = this.state;
@@ -288,7 +291,9 @@ const createBaseComponent = ({
         PluginComponent.sizeClassName || sizeClassName,
         PluginComponent.textWrapClassName || textWrapClassName,
         PluginComponent.customClassName,
-      ]).map(strategy => strategy(this.state.componentData, theme, this.styles, isMobile));
+      ]).map(strategy =>
+        strategy(this.state.componentData, theme, this.styles, isMobile, innerRCERenderedIn)
+      );
 
       const hasFocus = (isFocused ? !noPluginBorder : isPartOfSelection) && isEditorFocused;
 
@@ -296,6 +301,7 @@ const createBaseComponent = ({
         this.styles.pluginContainer,
         theme.pluginContainer,
         theme.pluginContainerWrapper,
+        withHorizontalScroll && this.styles.horizontalScrollbar,
         !isPartOfSelection && noPluginBorder && this.styles.noBorder,
         {
           [this.styles.pluginContainerMobile]: isMobile,
@@ -317,39 +323,15 @@ const createBaseComponent = ({
         isFocused && noPointerEventsOnFocus && this.styles.noPointerEventsOnFocus
       );
 
-      const sizeStyles = {
-        width: currentWidth || initialWidth,
-        height: currentHeight || initialHeight,
-        maxHeight: this.state.htmlPluginMaxHeight,
-      };
+      const sizeStyles = withHorizontalScroll
+        ? { position: 'unset' }
+        : {
+            width: currentWidth || initialWidth,
+            height: currentHeight || initialHeight,
+            maxHeight: this.state.htmlPluginMaxHeight,
+          };
 
       const component = (
-        <PluginComponent
-          {...this.props}
-          isMobile={isMobile}
-          settings={settings}
-          store={pubsub.store}
-          commonPubsub={commonPubsub}
-          theme={theme}
-          componentData={this.state.componentData}
-          componentState={this.state.componentState}
-          helpers={helpers}
-          t={t}
-          editorBounds={getEditorBounds()}
-          disableRightClick={disableRightClick}
-          anchorTarget={anchorTarget}
-          relValue={relValue}
-          locale={locale}
-          shouldRenderOptimizedImages={shouldRenderOptimizedImages}
-          iframeSandboxDomain={iframeSandboxDomain}
-          setInPluginEditingMode={setInPluginEditingMode}
-          getInPluginEditingMode={getInPluginEditingMode}
-          setComponentUrl={this.setComponentUrl}
-          renderInnerRCE={renderInnerRCE}
-        />
-      );
-
-      return (
         <div
           ref={this.containerRef}
           role="none"
@@ -360,7 +342,29 @@ const createBaseComponent = ({
           onContextMenu={this.handleContextMenu}
           {...decorationProps}
         >
-          {component}
+          <PluginComponent
+            {...this.props}
+            isMobile={isMobile}
+            settings={settings}
+            store={pubsub.store}
+            commonPubsub={commonPubsub}
+            theme={theme}
+            componentData={this.state.componentData}
+            componentState={this.state.componentState}
+            helpers={helpers}
+            t={t}
+            editorBounds={getEditorBounds()}
+            disableRightClick={disableRightClick}
+            anchorTarget={anchorTarget}
+            relValue={relValue}
+            locale={locale}
+            shouldRenderOptimizedImages={shouldRenderOptimizedImages}
+            iframeSandboxDomain={iframeSandboxDomain}
+            setInPluginEditingMode={setInPluginEditingMode}
+            getInPluginEditingMode={getInPluginEditingMode}
+            setComponentUrl={this.setComponentUrl}
+            renderInnerRCE={renderInnerRCE}
+          />
           <div
             role="none"
             data-hook={'componentOverlay'}
@@ -370,7 +374,14 @@ const createBaseComponent = ({
           />
         </div>
       );
-      /* eslint-enable jsx-a11y/anchor-has-content */
+
+      return withHorizontalScroll ? (
+        <div className={styles.horizontalScrollbarWrapper}>
+          <div className={styles.pluginWithHorizontalScrollbar}>{component}</div>
+        </div>
+      ) : (
+        component
+      );
     };
   };
 };